+++ ethpm_types/__init__.py
@@ -1,12 +1,15 @@
 from .abi import ABI
+from .ast import ASTNode
 from .base import BaseModel
 from .contract_type import Bytecode, ContractInstance, ContractType
 from .manifest import PackageManifest, PackageMeta
 from .source import Checksum, Compiler, Source
+from .sourcemap import PCMap, PCMapItem, SourceMap, SourceMapItem
 from .utils import HexBytes
 
 __all__ = [
     "ABI",
+    "ASTNode",
     "BaseModel",
     "Bytecode",
     "Checksum",
@@ -16,5 +19,9 @@ __all__ = [
     "HexBytes",
     "PackageMeta",
     "PackageManifest",
+    "PCMap",
+    "PCMapItem",
     "Source",
+    "SourceMap",
+    "SourceMapItem",
 ]
