+++ pyproject.toml
@@ -36,4 +36,5 @@ use_parentheses = true
 
 [tool.mypy]
 exclude = "build"
+check_untyped_defs = false
 plugins = ["pydantic.mypy"]
